add get requests

+++ cms/src/main/java/dia/kids/cms/repository/PictureRepository.java
@@ -3,6 +3,6 @@ package dia.kids.cms.repository;
 import dia.kids.cms.model.Picture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PictureRepository extends JpaRepository<Picture, Long> {
+public interface PictureRepository extends JpaRepository<Picture, Integer> {
 
 }
